Move message component classes into their own folder
