+++ Irene/Commands/Roles.cs
@@ -3,6 +3,7 @@ using System.IO;
 
 using DSharpPlus.Entities;
 
+using Irene.Components;
 using static Irene.Program;
 
 namespace Irene.Commands {
